Refactor mock_download_manager
