+++ src/datasets/download/mock_download_manager.py
@@ -21,9 +21,9 @@ import urllib.parse
 from pathlib import Path
 from typing import Callable, List, Optional, Union
 
-from .file_utils import cached_path, hf_github_url
-from .logging import get_logger
-from .version import Version
+from datasets.utils.file_utils import cached_path, hf_github_url
+from datasets.utils.logging import get_logger
+from datasets.utils.version import Version
 
 
 logger = get_logger(__name__)
